backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sympy>=1.13
pydantic>=2.8
anyio>=4.0
numpy>=1.26
groq>=0.9
python-dotenv>=1.0
